Changes from #298
